requirements.txt
django==3.2.11
djangorestframework==3.13.1
django-cors-headers
psycopg2
django-import-export
django-payu-latam